fix: ボタンの装飾を微修正 #134
