+++ frontend/src/components/models/task/TaskEditModal.tsx
@@ -7,7 +7,9 @@ import { SearchMemberField } from 'components/ui/form/SearchMemberField'
 import { TaskEditTitleField } from 'components/models/task/TaskEditTitleField'
 import { TaskEditOverviewField } from 'components/models/task/TaskEditOverviewField'
 import { TaskEditStatusPointField } from 'components/models/task/TaskEditStatusPointField'
+import { CoarseButton } from 'components/ui/button/CoarseButton'
 import { convertIntoRGBA } from 'utils/color/convertIntoRGBA'
+import { strokeTextShadow } from 'utils/strokeTextShadow'
 import { calculateMinSizeBasedOnFigma } from 'utils/calculateSizeBasedOnFigma'
 import { useTaskEndDateEditForm } from 'hooks/useTaskEndDateEditForm'
 import { useTaskUserSelectForm } from 'hooks/useTaskUserSelectForm'
@@ -58,7 +60,7 @@ export const TaskEditModal: FCX<Props> = ({
           onChange={onChange}
         />
         <StyledSearchMemberField setUserDatas={setUserDatas} userDatas={userDatas} />
-        <TaskEditStatusPointField
+        <StyledTaskEditStatusPointField
           id={id}
           technology={technology}
           solution={solution}
@@ -67,6 +69,7 @@ export const TaskEditModal: FCX<Props> = ({
           design={design}
           plan={plan}
         />
+        <StyledDeleteButton text="タスクを削除" disabled={false} />
       </StyledRightColumn>
     </StyledModal>
   )
@@ -129,3 +132,37 @@ const StyledCalenderField = styled(CalenderField)`
   margin-bottom: ${calculateMinSizeBasedOnFigma(19)};
 `
 const StyledSearchMemberField = StyledCalenderField.withComponent(SearchMemberField)
+const StyledTaskEditStatusPointField = styled(TaskEditStatusPointField)`
+  margin-bottom: ${calculateMinSizeBasedOnFigma(30)};
+`
+type Disabled = { disabled: boolean }
+const StyledDeleteButton = styled(CoarseButton).attrs<Disabled>(({ disabled }) => ({
+  disabled,
+}))<Disabled>`
+  width: 100%;
+  height: ${calculateMinSizeBasedOnFigma(32)};
+  ${({ disabled, theme }) => {
+    if (disabled) {
+      return css`
+        color: ${theme.COLORS.SILVER};
+        > div {
+          background-color: ${convertIntoRGBA(theme.COLORS.ALTO, 0.55)};
+          > div > div {
+            background-color: ${convertIntoRGBA(theme.COLORS.NOBEL, 0.64)};
+          }
+        }
+      `
+    } else {
+      return css`
+        ${strokeTextShadow('1px', theme.COLORS.MONDO)};
+        color: ${theme.COLORS.WHITE};
+        > div {
+          background-color: ${convertIntoRGBA(theme.COLORS.TEMPTRESS, 0.2)};
+          > div > div {
+            background-color: ${convertIntoRGBA(theme.COLORS.RED_BERRY, 0.6)};
+          }
+        }
+      `
+    }
+  }}
+`
