+++ frontend/src/components/models/task/TaskEditStatusPointField.tsx
@@ -1,10 +1,9 @@
 import React, { FCX } from 'react'
 import styled, { css } from 'styled-components'
 import { TaskStatusPointField } from 'components/models/task/TaskStatusPointField'
-import { CoarseButton } from 'components/ui/button/CoarseButton'
+import { CoarseRedOxideButton } from 'components/ui/button/CoarseRedOxideButton'
 import { useTaskStatusPointEditForm } from 'hooks/useTaskStatusPointEditForm'
 import { calculateMinSizeBasedOnFigma } from 'utils/calculateSizeBasedOnFigma'
-import { convertIntoRGBA } from 'utils/color/convertIntoRGBA'
 import { STATUS_TYPE } from 'consts/status'
 import { StatusParam } from 'types/status'
 
@@ -30,7 +29,7 @@ export const TaskEditStatusPointField: FCX<Props> = ({ className, id, ...initial
         />
       ))}
       <StyledSaveButtonWrapper>
-        <StyledSaveButton text="保存" onClick={onClickSaveButton} disabled={disabled} />
+        <CoarseRedOxideButton text="保存" onClick={onClickSaveButton} disabled={disabled} />
       </StyledSaveButtonWrapper>
     </StyledWrapper>
   )
@@ -55,34 +54,4 @@ const StyledSaveButtonWrapper = styled.div`
   justify-content: flex-end;
   align-items: center;
   width: 100%;
-`
-type Disabled = { disabled: boolean }
-const StyledSaveButton = styled(CoarseButton).attrs<Disabled>(({ disabled }) => ({
-  disabled,
-}))<Disabled>`
-  width: ${calculateMinSizeBasedOnFigma(64)};
-  height: ${calculateMinSizeBasedOnFigma(32)};
-  ${({ disabled, theme }) => {
-    if (disabled) {
-      return css`
-        color: ${theme.COLORS.SILVER};
-        > div {
-          background-color: ${convertIntoRGBA(theme.COLORS.ALTO, 0.55)};
-          > div > div {
-            background-color: ${convertIntoRGBA(theme.COLORS.NOBEL, 0.64)};
-          }
-        }
-      `
-    } else {
-      return css`
-        color: ${theme.COLORS.BRANDY};
-        > div {
-          background-color: ${convertIntoRGBA(theme.COLORS.TEMPTRESS, 0.2)};
-          > div > div {
-            background-color: ${convertIntoRGBA(theme.COLORS.RED_OXIDE, 0.45)};
-          }
-        }
-      `
-    }
-  }}
 `
